requirements.txt
google-generativeai>=0.5.0
faiss-cpu>=1.7.4
pdfplumber>=0.10.2
python-docx>=1.1.0
python-dotenv>=1.0.1
openpyxl>=3.1.2
streamlit>=1.32.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
pydantic>=2.5.0
sentence-transformers>=2.7.0

